refactor: 코드 스타일 개선 및 불필요한 형 변환 제거

+++ EngineSIU/EngineSIU/Engine/Source/Editor/PropertyEditor/ControlEditorPanel.cpp
@@ -468,7 +468,7 @@ void ControlEditorPanel::CreateSRTButton(ImVec2 ButtonSize) const
 
 uint64 ControlEditorPanel::ConvertSelectionToFlags(const bool selected[]) const
 {
-    uint64 flags = static_cast<uint64>(EEngineShowFlags::None);
+    uint64 flags = EEngineShowFlags::None;
 
     if (selected[0])
         flags |= static_cast<uint64>(EEngineShowFlags::SF_AABB);
